+++ ontolutils/ex/dcat/__init__.py
@@ -1,3 +1,3 @@
-from .resource import Dataset, Distribution, Resource, DataService
+from .resource import Dataset, Distribution, Resource, DataService, __version__
 
-__all__ = ['Dataset', 'Distribution', 'Resource', 'DataService']
+__all__ = ['Dataset', 'Distribution', 'Resource', 'DataService', '__version__']