+++ ontolutils/ex/foaf/__init__.py
@@ -1,7 +1,8 @@
-from .agent import Agent, Person, Organization
+from .agent import Agent, Person, Organization, __version__
 
 __all__ = [
     "Agent",
     "Person",
     "Organization",
+    "__version__",
 ]