+++ ontolutils/ex/hdf5/__init__.py
@@ -1,3 +1,3 @@
-from .allotrope import File, Group, Dataset
+from .allotrope import File, Group, Dataset, __version__
 
-__all__ = ("File", "Group", "Dataset")
+__all__ = ("File", "Group", "Dataset", "__version__")