+++ ontolutils/ex/m4i/__init__.py
@@ -1,8 +1,9 @@
-from .m4i import Method, TextVariable, NumericalVariable, ProcessingStep, Tool
+from .m4i import Method, TextVariable, NumericalVariable, ProcessingStep, Tool, __version__
 
 __all__ = ('ProcessingStep',
            'Method',
            'Tool',
            'NumericalVariable',
-           'TextVariable'
+           'TextVariable',
+           '__version__'
            )