+++ ontolutils/ex/prov/__init__.py
@@ -1,5 +1,8 @@
 from .attribution import Attribution, Person, Organization, Agent, SoftwareAgent
 from .activity import Activity
+
+__version__ = "2013.19.05"
+
 __all__ = [
     "Attribution",
     "Person",
@@ -7,4 +10,5 @@ __all__ = [
     "Agent",
     "SoftwareAgent",
     "Activity",
+    "__version__"
 ]