+++ ontolutils/ex/spdx/__init__.py
@@ -1,3 +1,4 @@
 from .spdx import Checksum
 
-__all__ = ['Checksum']
+__version__ = "2.3"
+__all__ = ['Checksum', '__version__']